split container
