+++ renderer/features/toolbar/EditorToolbar.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import * as O from 'fp-ts/Option';
 import { Box, IconButton } from '@material-ui/core';
 import { TextFields, PhotoLibrary, Crop32 } from '@material-ui/icons';
-import { ShapeContainer } from '../../store-unstated';
+import { ShapeContainer } from '../../store/ShapesContainer';
 
 const EditorToolbar = (): React.ReactElement => {
   const shapes = ShapeContainer.useContainer();
